Initial pass in removing more code from individual calls.  They now know how to populate existing get/post param dictionaries in the superclass WebServiceParameters and Parse a *text* response into the appropriate format.  This leads to quick calls being written like login.Parse(makeCallUsingAService(login)) which seems a bit odd, but this decouples execution flow, which is important for any asynchronous running which would be more natural along the lines of IWebService service = new AsyncService(login); yield return service; login.Parse(service.text) //gimme a result!
Also it helps in testing independently the result handling from the network calls.  I’m not sure they should be instance methods, but this is a step in a direction that seems appropriately flexible.

put some thought in handling synchronous flow more seamlessly in the web service response, but left it commented out for now.