do guiLayout so that it works with transfluent
make it so the guywrapper isn’t automatically imported by other scripts by moving the namespace from transfluent to tranfluent.guiwrapper .

sample of “converted” internaltionaltextdisplaywithtransfluent
NOTE: default parameters do not seem to be working as expected :(  Not sure how to detect those.  Decompiled source does not indicate default values for these.  Perhaps that is the wrong field?  Or is it just that GUILayout only can be null as a parameter, and does not seem to be explicitly stated as that in the documentation?  The manually created version of the GUI class has this problem too.